Move DicomText to separate file
